dtar: refactor archive create/extract into files and functions for mfu
