+++ experimental/Makefile.am
@@ -1,2 +1,2 @@
 #SUBDIRS = dgrep dparallel dtar dfind dsh
-SUBDIRS = dparallel dsh dbz2
+SUBDIRS = dparallel dsh dbz2 dtar
