+++ experimental/dtar/dtar.c
@@ -26,12 +26,6 @@
  *
  */
 
-#ifndef _GNU_SOURCE
-#define _GNU_SOURCE
-#endif
-
-#define _LARGEFILE64_SOURCE
-
 #include <errno.h>
 #include <stdio.h>
 #include <stdbool.h>
@@ -47,12 +41,7 @@
 #include <getopt.h>
 
 #include "mfu.h"
-
-#define DTAR_HDR_LENGTH 1536
-
-typedef enum {
-    COPY_DATA
-} DTAR_operation_code_t;
+#include "mfu_flist_archive.h"
 
 /* common structures */
 
@@ -62,33 +51,7 @@ typedef struct {
     char*   dest_path;
     bool    preserve;
     int     flags;
-} DTAR_options_t;
-
-typedef struct {
-    const char* name;
-    int fd_tar;
-    int flags;
-} DTAR_writer_t;
-
-typedef struct {
-    uint64_t total_dirs;
-    uint64_t total_files;
-    uint64_t total_links;
-    uint64_t total_size;
-    uint64_t total_bytes_copied;
-    double  wtime_started;
-    double  wtime_ended;
-    time_t  time_started;
-    time_t  time_ended;
-} DTAR_statistics_t;
-
-typedef struct {
-    uint64_t file_size;
-    uint64_t chunk_index;
-    uint64_t offset;
-    DTAR_operation_code_t code;
-    char* operand;
-} DTAR_operation_t;
+} mfu_flist_archive_options_t;
 
 /* The opts_blocksize option is optional and is used to specify
  * a blocksize for compression.
@@ -108,10 +71,6 @@ static size_t  opts_chunksize = 1024 * 1024;
 static int     opts_blocksize = 9;
 static ssize_t opts_memory    = -1;
 
-mfu_param_path* src_params;
-mfu_param_path dest_param;
-int num_src_params;
-
 static void DTAR_abort(int code)
 {
     MPI_Abort(MPI_COMM_WORLD, code);
@@ -123,547 +82,6 @@ static void DTAR_exit(int code)
     mfu_finalize();
     MPI_Finalize();
     exit(code);
-}
-
-mfu_flist DTAR_flist;
-uint64_t* DTAR_offsets = NULL;
-int DTAR_global_rank;
-DTAR_options_t DTAR_user_opts;
-DTAR_writer_t DTAR_writer;
-DTAR_statistics_t DTAR_statistics;
-uint64_t DTAR_count = 0;
-int DTAR_rank, DTAR_size;
-
-static void DTAR_write_header(struct archive* ar, uint64_t idx, uint64_t offset)
-{
-    /* allocate and entry for this item */
-    struct archive_entry* entry = archive_entry_new();
-
-    /* get file name for this item */
-    /* fill up entry, FIXME: the uglyness of removing leading slash */
-    const char* fname = mfu_flist_file_get_name(DTAR_flist, idx);
-    archive_entry_copy_pathname(entry, &fname[1]);
-
-    if (DTAR_user_opts.preserve) {
-        struct archive* source = archive_read_disk_new();
-        archive_read_disk_set_standard_lookup(source);
-        int fd = open(fname, O_RDONLY);
-        if (archive_read_disk_entry_from_file(source, entry, fd, NULL) != ARCHIVE_OK) {
-            MFU_LOG(MFU_LOG_ERR, "archive_read_disk_entry_from_file(): %s", archive_error_string(ar));
-        }
-        archive_read_free(source);
-        close(fd);
-    } else {
-        /* TODO: read stat info from mfu_flist */
-        struct stat stbuf;
-        mfu_lstat(fname, &stbuf);
-        archive_entry_copy_stat(entry, &stbuf);
-
-        /* set user name of owner */
-        const char* uname = mfu_flist_file_get_username(DTAR_flist, idx);
-        archive_entry_set_uname(entry, uname);
-
-        /* set group name */
-        const char* gname = mfu_flist_file_get_groupname(DTAR_flist, idx);
-        archive_entry_set_gname(entry, gname);
-    }
-
-    /* TODO: Seems to be a bug here potentially leading to corrupted
-     * archive files.  archive_write_free also writes two blocks of
-     * NULL bytes at the end of an archive file, however, each rank
-     * will have a different view of the length of the file, so one
-     * rank may write its NULL blocks over top of the actual data
-     * written by another rank */
-
-    /* write entry info to archive */
-    struct archive* dest = archive_write_new();
-    archive_write_set_format_pax(dest);
-
-    if (archive_write_open_fd(dest, DTAR_writer.fd_tar) != ARCHIVE_OK) {
-        MFU_LOG(MFU_LOG_ERR, "archive_write_open_fd(): %s", archive_error_string(ar));
-    }
-
-    /* seek to offset in tar archive for this file */
-    lseek(DTAR_writer.fd_tar, offset, SEEK_SET);
-
-    /* write header for this item */
-    if (archive_write_header(dest, entry) != ARCHIVE_OK) {
-        MFU_LOG(MFU_LOG_ERR, "archive_write_header(): %s", archive_error_string(ar));
-    }
-
-    archive_entry_free(entry);
-    archive_write_free(dest);
-}
-
-static char* DTAR_encode_operation(DTAR_operation_code_t code, const char* operand,
-                            uint64_t fsize, uint64_t chunk_idx, uint64_t offset)
-{
-    size_t opsize = (size_t) CIRCLE_MAX_STRING_LEN;
-    char* op = (char*) MFU_MALLOC(opsize);
-    size_t len = strlen(operand);
-
-    int written = snprintf(op, opsize,
-                           "%" PRIu64 ":%" PRIu64 ":%" PRIu64 ":%d:%d:%s",
-                           fsize, chunk_idx, offset, code, (int) len, operand);
-
-    if (written >= opsize) {
-        MFU_LOG(MFU_LOG_ERR, "Exceed libcirlce message size");
-        DTAR_abort(EXIT_FAILURE);
-    }
-
-    return op;
-}
-
-static DTAR_operation_t* DTAR_decode_operation(char* op)
-{
-    DTAR_operation_t* ret = (DTAR_operation_t*) MFU_MALLOC(
-                                sizeof(DTAR_operation_t));
-
-    if (sscanf(strtok(op, ":"), "%" SCNu64, &(ret->file_size)) != 1) {
-        MFU_LOG(MFU_LOG_ERR, "Could not decode file size attribute.");
-        DTAR_abort(EXIT_FAILURE);
-    }
-
-    if (sscanf(strtok(NULL, ":"), "%" SCNu64, &(ret->chunk_index)) != 1) {
-        MFU_LOG(MFU_LOG_ERR, "Could not decode chunk index attribute.");
-        DTAR_abort(EXIT_FAILURE);
-    }
-
-    if (sscanf(strtok(NULL, ":"), "%" SCNu64, &(ret->offset)) != 1) {
-        MFU_LOG(MFU_LOG_ERR, "Could not decode source base offset attribute.");
-        DTAR_abort(EXIT_FAILURE);
-    }
-
-    if (sscanf(strtok(NULL, ":"), "%d", (int*) & (ret->code)) != 1) {
-        MFU_LOG(MFU_LOG_ERR, "Could not decode stage code attribute.");
-        DTAR_abort(EXIT_FAILURE);
-    }
-
-    /* get number of characters in operand string */
-    int op_len;
-    char* str = strtok(NULL, ":");
-    if (sscanf(str, "%d", &op_len) != 1) {
-        MFU_LOG(MFU_LOG_ERR, "Could not decode operand string length.");
-        DTAR_abort(EXIT_FAILURE);
-    }
-
-    /* skip over digits and trailing ':' to get pointer to operand */
-    char* operand = str + strlen(str) + 1;
-    operand[op_len] = '\0';
-    ret->operand = operand;
-
-    return ret;
-}
-
-static void DTAR_enqueue_copy(CIRCLE_handle* handle)
-{
-    for (uint64_t idx = 0; idx < DTAR_count; idx++) {
-        /* add copy work only for files */
-        mfu_filetype type = mfu_flist_file_get_type(DTAR_flist, idx);
-        if (type == MFU_TYPE_FILE) {
-            /* get name and size of file */
-            const char* name = mfu_flist_file_get_name(DTAR_flist, idx);
-            uint64_t size = mfu_flist_file_get_size(DTAR_flist, idx);
-
-            /* compute offset for first byte of file content */
-            uint64_t dataoffset = DTAR_offsets[idx] + DTAR_HDR_LENGTH;
-
-            /* compute number of chunks */
-            uint64_t num_chunks = size / DTAR_user_opts.chunk_size;
-            for (uint64_t chunk_idx = 0; chunk_idx < num_chunks; chunk_idx++) {
-                char* newop = DTAR_encode_operation(
-                                  COPY_DATA, name, size, chunk_idx, dataoffset);
-                handle->enqueue(newop);
-                mfu_free(&newop);
-            }
-
-            /* create copy work for possibly last item */
-            if (num_chunks * DTAR_user_opts.chunk_size < size || num_chunks == 0) {
-                char* newop = DTAR_encode_operation(
-                                  COPY_DATA, name, size, num_chunks, dataoffset);
-                handle->enqueue(newop);
-                mfu_free(&newop);
-            }
-        }
-    }
-}
-
-static void DTAR_perform_copy(CIRCLE_handle* handle)
-{
-    char opstr[CIRCLE_MAX_STRING_LEN];
-    char iobuf[FD_BLOCK_SIZE];
-
-    int out_fd = DTAR_writer.fd_tar;
-
-    handle->dequeue(opstr);
-    DTAR_operation_t* op = DTAR_decode_operation(opstr);
-
-    uint64_t in_offset = DTAR_user_opts.chunk_size * op->chunk_index;
-    int in_fd = open(op->operand, O_RDONLY);
-
-    ssize_t num_of_bytes_read = 0;
-    ssize_t num_of_bytes_written = 0;
-    ssize_t total_bytes_written = 0;
-
-    uint64_t out_offset = op->offset + in_offset;
-
-    lseek(in_fd, in_offset, SEEK_SET);
-    lseek(out_fd, out_offset, SEEK_SET);
-
-    while (total_bytes_written < DTAR_user_opts.chunk_size) {
-        num_of_bytes_read = read(in_fd, &iobuf[0], sizeof(iobuf));
-        if (!num_of_bytes_read) { break; }
-        num_of_bytes_written = write(out_fd, &iobuf[0], num_of_bytes_read);
-        total_bytes_written += num_of_bytes_written;
-    }
-
-    uint64_t num_chunks = op->file_size / DTAR_user_opts.chunk_size;
-    uint64_t rem = op->file_size - DTAR_user_opts.chunk_size * num_chunks;
-    uint64_t last_chunk = (rem) ? num_chunks : num_chunks - 1;
-
-    /* handle last chunk */
-    if (op->chunk_index == last_chunk) {
-        int padding = 512 - (int) (op->file_size % 512);
-        if (padding > 0 && padding != 512) {
-            char* buff = (char*) calloc(padding, sizeof(char));
-            write(out_fd, buff, padding);
-        }
-    }
-
-    close(in_fd);
-    mfu_free(&op);
-}
-
-static void DTAR_epilogue(void)
-{
-    double rel_time = DTAR_statistics.wtime_ended - \
-                      DTAR_statistics.wtime_started;
-    if (DTAR_rank == 0) {
-        char starttime_str[256];
-        struct tm* localstart = localtime(&(DTAR_statistics.time_started));
-        strftime(starttime_str, 256, "%b-%d-%Y, %H:%M:%S", localstart);
-
-        char endtime_str[256];
-        struct tm* localend = localtime(&(DTAR_statistics.time_ended));
-        strftime(endtime_str, 256, "%b-%d-%Y, %H:%M:%S", localend);
-
-        /* add two 512 blocks at the end */
-        DTAR_statistics.total_size += 512 * 2;
-
-        /* convert bandwidth to unit */
-        double agg_rate_tmp;
-        double agg_rate = (double) DTAR_statistics.total_size / rel_time;
-        const char* agg_rate_units;
-        mfu_format_bytes(agg_rate, &agg_rate_tmp, &agg_rate_units);
-
-        MFU_LOG(MFU_LOG_INFO, "Started:    %s", starttime_str);
-        MFU_LOG(MFU_LOG_INFO, "Completed:  %s", endtime_str);
-        MFU_LOG(MFU_LOG_INFO, "Total archive size: %" PRIu64, DTAR_statistics.total_size);
-        MFU_LOG(MFU_LOG_INFO, "Rate: %.3lf %s " \
-                "(%.3" PRIu64 " bytes in %.3lf seconds)", \
-                agg_rate_tmp, agg_rate_units, DTAR_statistics.total_size, rel_time);
-    }
-}
-
-static void mfu_flist_archive_create_libcircle(mfu_flist flist, const char* archivefile)
-{
-    DTAR_flist = flist;
-
-    /* TODO: stripe the archive file if on parallel file system */
-
-    /* create the archive file */
-    DTAR_writer.name = archivefile;
-    DTAR_writer.flags = O_WRONLY | O_CREAT | O_CLOEXEC | O_LARGEFILE;
-    DTAR_writer.fd_tar = open(archivefile, DTAR_writer.flags, 0664);
-
-    /* get number of items in our portion of the list */
-    DTAR_count = mfu_flist_size(DTAR_flist);
-
-    /* allocate memory for file sizes and offsets */
-    uint64_t* fsizes = (uint64_t*) MFU_MALLOC(DTAR_count * sizeof(uint64_t));
-    DTAR_offsets     = (uint64_t*) MFU_MALLOC(DTAR_count * sizeof(uint64_t));
-
-    /* compute local offsets for each item and total
-     * bytes we're contributing to the archive */
-    uint64_t idx;
-    uint64_t offset = 0;
-    for (idx = 0; idx < DTAR_count; idx++) {
-        /* assume the item takes no space */
-        fsizes[idx] = 0;
-
-        /* identify item type to compute its size in the archive */
-        mfu_filetype type = mfu_flist_file_get_type(DTAR_flist, idx);
-        if (type == MFU_TYPE_DIR || type == MFU_TYPE_LINK) {
-            /* directories and symlinks only need the header */
-            fsizes[idx] = DTAR_HDR_LENGTH;
-        } else if (type == MFU_TYPE_FILE) {
-            /* regular file requires a header, plus file content,
-             * and things are packed into blocks of 512 bytes */
-            uint64_t fsize = mfu_flist_file_get_size(DTAR_flist, idx);
-
-            /* determine whether file size is integer multiple of 512 bytes */
-            uint64_t rem = fsize % 512;
-            if (rem == 0) {
-                /* file content is multiple of 512 bytes, so perfect fit */
-                fsizes[idx] = fsize + DTAR_HDR_LENGTH;
-            } else {
-                /* TODO: check and explain this math */
-                fsizes[idx] = (fsize / 512 + 4) * 512;
-            }
-
-        }
-
-        /* increment our local offset for this item */
-        DTAR_offsets[idx] = offset;
-        offset += fsizes[idx];
-    }
-
-    /* execute scan to figure our global base offset in the archive file */
-    uint64_t global_offset = 0;
-    MPI_Scan(&offset, &global_offset, 1, MPI_UINT64_T, MPI_SUM, MPI_COMM_WORLD);
-    global_offset -= offset;
-
-    /* update offsets for each of our file to their global offset */
-    for (idx = 0; idx < DTAR_count; idx++) {
-        DTAR_offsets[idx] += global_offset;
-    }
-
-    /* create an archive */
-    struct archive* ar = archive_write_new();
-
-    archive_write_set_format_pax(ar);
-
-    int r = archive_write_open_fd(ar, DTAR_writer.fd_tar);
-    if (r != ARCHIVE_OK) {
-        MFU_LOG(MFU_LOG_ERR, "archive_write_open_fd(): %s", archive_error_string(ar));
-        DTAR_abort(EXIT_FAILURE);
-    }
-
-    /* write headers for our files */
-    for (idx = 0; idx < DTAR_count; idx++) {
-        mfu_filetype type = mfu_flist_file_get_type(DTAR_flist, idx);
-        if (type == MFU_TYPE_FILE || type == MFU_TYPE_DIR || type == MFU_TYPE_LINK) {
-            DTAR_write_header(ar, idx, DTAR_offsets[idx]);
-        }
-    }
-
-    /* prepare libcircle */
-    DTAR_global_rank = CIRCLE_init(0, NULL, CIRCLE_SPLIT_EQUAL | CIRCLE_CREATE_GLOBAL);
-    CIRCLE_loglevel loglevel = CIRCLE_LOG_WARN;
-    CIRCLE_enable_logging(loglevel);
-
-    /* register callbacks */
-    CIRCLE_cb_create(&DTAR_enqueue_copy);
-    CIRCLE_cb_process(&DTAR_perform_copy);
-
-    /* run the libcircle job to copy data into archive file */
-    CIRCLE_begin();
-    CIRCLE_finalize();
-
-    /* compute total bytes copied */
-    uint64_t archive_size = 0;
-    MPI_Allreduce(&offset, &archive_size, 1, MPI_UINT64_T, MPI_SUM, MPI_COMM_WORLD);
-    DTAR_statistics.total_size = archive_size;
-
-    /* clean up */
-    mfu_free(&fsizes);
-    mfu_free(&DTAR_offsets);
-
-    /* close archive file */
-    archive_write_free(ar);
-    mfu_close(DTAR_writer.name, DTAR_writer.fd_tar);
-}
-
-static void errmsg(const char* m)
-{
-    fprintf(stderr, "%s\n", m);
-}
-
-static void msg(const char* m)
-{
-    fprintf(stdout, "%s", m);
-}
-
-static int copy_data(struct archive* ar, struct archive* aw)
-{
-    const void* buff;
-    size_t size;
-    off_t offset;
-    for (;;) {
-        int r = archive_read_data_block(ar, &buff, &size, &offset);
-        if (r == ARCHIVE_EOF) {
-            return ARCHIVE_OK;
-        }
-        if (r != ARCHIVE_OK) {
-            return r;
-        }
-
-        r = archive_write_data_block(aw, buff, size, offset);
-        if (r != ARCHIVE_OK) {
-            errmsg(archive_error_string(ar));
-            return r;
-        }
-    }
-    return 0;
-}
-
-static void extract_archive(const char* filename, bool verbose, int flags)
-{
-    int r;
-
-    /* TODO: this needs to be parallelized */
-
-    /* initiate archive object for reading */
-    struct archive* a = archive_read_new();
-
-    /* initiate archive object for writing */
-    struct archive* ext = archive_write_disk_new();
-    archive_write_disk_set_options(ext, flags);
-
-    /* we want all the format supports */
-    archive_read_support_filter_bzip2(a);
-    archive_read_support_filter_gzip(a);
-    archive_read_support_filter_compress(a);
-    archive_read_support_format_tar(a);
-
-    archive_write_disk_set_standard_lookup(ext);
-
-    if (filename != NULL && strcmp(filename, "-") == 0) {
-        filename = NULL;
-    }
-
-    /* blocksize set to 1024K */
-    if ((r = archive_read_open_filename(a, filename, 10240))) {
-        errmsg(archive_error_string(a));
-        exit(r);
-    }
-
-    struct archive_entry* entry;
-    for (;;) {
-        r = archive_read_next_header(a, &entry);
-        if (r == ARCHIVE_EOF) {
-            break;
-        }
-        if (r != ARCHIVE_OK) {
-            errmsg(archive_error_string(a));
-            exit(r);
-        }
-
-        if (verbose) {
-            msg("x ");
-        }
-
-        if (verbose) {
-            msg(archive_entry_pathname(entry));
-        }
-
-        r = archive_write_header(ext, entry);
-        if (r != ARCHIVE_OK) {
-            errmsg(archive_error_string(a));
-        } else {
-            copy_data(a, ext);
-        }
-
-        if (verbose) {
-            msg("\n");
-        }
-    }
-
-    archive_read_close(a);
-    archive_read_free(a);
-}
-
-static void DTAR_check_paths(int numparams, mfu_param_path* srcparams, mfu_param_path destparam)
-{
-    int valid = 1;
-    int i;
-    int num_readable = 0;
-    for (i = 0; i < numparams; i++) {
-        char* path = srcparams[i].path;
-        if (mfu_access(path, R_OK) == 0) {
-            num_readable++;
-        } else {
-            /* not readable */
-            char* orig = srcparams[i].orig;
-            MFU_LOG(MFU_LOG_ERR, "Could not read '%s' errno=%d %s",
-                    orig, errno, strerror(errno));
-        }
-    }
-
-    /* verify we have at least one valid source */
-    if (num_readable < 1) {
-        MFU_LOG(MFU_LOG_ERR, "At least one valid source must be specified");
-        valid = 0;
-        goto bcast;
-    }
-
-    /* check destination */
-    if (destparam.path_stat_valid) {
-        if (DTAR_rank == 0) {
-            MFU_LOG(MFU_LOG_WARN, "Destination target exists, we will overwrite");
-        }
-    } else {
-        /* compute path to parent of destination archive */
-        mfu_path* parent = mfu_path_from_str(destparam.path);
-        mfu_path_dirname(parent);
-        char* parent_str = mfu_path_strdup(parent);
-        mfu_path_delete(&parent);
-
-        /* check if parent is writable */
-        if (mfu_access(parent_str, W_OK) < 0) {
-            MFU_LOG(MFU_LOG_ERR, "Destination parent directory is not wriable: '%s' ",
-                    parent_str);
-            valid = 0;
-            mfu_free(&parent_str);
-            goto bcast;
-        }
-
-        mfu_free(&parent_str);
-    }
-
-    /* at this point, we know
-     * (1) destination doesn't exist
-     * (2) parent directory is writable
-     */
-
-bcast:
-    MPI_Bcast(&valid, 1, MPI_INT, 0, MPI_COMM_WORLD);
-
-    if (! valid) {
-        if (DTAR_global_rank == 0) {
-            MFU_LOG(MFU_LOG_ERR, "Exiting run");
-        }
-        MPI_Barrier(MPI_COMM_WORLD);
-        DTAR_exit(EXIT_FAILURE);
-    }
-}
-
-static void DTAR_parse_path_args(int numpaths, const char** pathlist, const char* dstfile)
-{
-    if (pathlist == NULL || numpaths < 1) {
-        if (DTAR_global_rank == 0) {
-            fprintf(stderr, "\nYou must provide at least one source file or directory\n");
-            DTAR_exit(EXIT_FAILURE);
-        }
-    }
-
-    /* allocate space to record info for each source */
-    src_params = NULL;
-    num_src_params = numpaths;
-    size_t src_params_bytes = ((size_t) num_src_params) * sizeof(mfu_param_path);
-    src_params = (mfu_param_path*) MFU_MALLOC(src_params_bytes);
-
-    /* process each source path */
-    mfu_param_path_set_all(numpaths, pathlist, src_params);
-
-    /* standardize destination path */
-    mfu_param_path_set(dstfile, &dest_param);
-
-    /* copy destination to user opts structure */
-    DTAR_user_opts.dest_path = MFU_STRDUP(dest_param.path);
-
-    /* check that source and destination are okay */
-    DTAR_check_paths(num_src_params, src_params, dest_param);
 }
 
 static void print_usage(void)
@@ -688,13 +106,14 @@ static void print_usage(void)
     return;
 }
 
-int main(int argc, const char** argv)
+int main(int argc, char** argv)
 {
     MPI_Init(&argc, &argv);
     mfu_init();
 
-    MPI_Comm_rank(MPI_COMM_WORLD, &DTAR_rank);
-    MPI_Comm_size(MPI_COMM_WORLD, &DTAR_size);
+    int rank, size;
+    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+    MPI_Comm_size(MPI_COMM_WORLD, &size);
 
     int option_index = 0;
     static struct option long_options[] = {
@@ -764,7 +183,7 @@ int main(int argc, const char** argv)
                 usage = 1;
                 break;
             default:
-                if (DTAR_rank == 0) {
+                if (rank == 0) {
                     printf("?? getopt returned character code 0%o ??\n", c);
                 }
                 usage = 1;
@@ -773,7 +192,7 @@ int main(int argc, const char** argv)
 
     /* print usage if we need to */
     if (usage) {
-        if (DTAR_rank == 0) {
+        if (rank == 0) {
             print_usage();
         }
         mfu_finalize();
@@ -789,12 +208,12 @@ int main(int argc, const char** argv)
         mfu_debug_level = MFU_LOG_ERR;
     }
 
-    if (!opts_create && !opts_extract && DTAR_global_rank == 0) {
+    if (!opts_create && !opts_extract && rank == 0) {
         MFU_LOG(MFU_LOG_ERR, "One of extract(x) or create(c) need to be specified");
         DTAR_exit(EXIT_FAILURE);
     }
 
-    if (opts_create && opts_extract && DTAR_global_rank == 0) {
+    if (opts_create && opts_extract && rank == 0) {
         MFU_LOG(MFU_LOG_ERR, "Only one of extraction(x) or create(c) can be specified");
         DTAR_exit(EXIT_FAILURE);
     }
@@ -806,50 +225,54 @@ int main(int argc, const char** argv)
     }
 
     /* done by default */
-    DTAR_user_opts.flags  = ARCHIVE_EXTRACT_TIME;
-    DTAR_user_opts.flags |= ARCHIVE_EXTRACT_OWNER;
-    DTAR_user_opts.flags |= ARCHIVE_EXTRACT_PERM;
-    DTAR_user_opts.flags |= ARCHIVE_EXTRACT_ACL;
-    DTAR_user_opts.flags |= ARCHIVE_EXTRACT_FFLAGS;
+    mfu_archive_options_t archive_opts;
+    archive_opts.flags  = ARCHIVE_EXTRACT_TIME;
+    archive_opts.flags |= ARCHIVE_EXTRACT_OWNER;
+    archive_opts.flags |= ARCHIVE_EXTRACT_PERM;
+    archive_opts.flags |= ARCHIVE_EXTRACT_ACL;
+    archive_opts.flags |= ARCHIVE_EXTRACT_FFLAGS;
 
     if (opts_preserve) {
-        DTAR_user_opts.flags |= ARCHIVE_EXTRACT_XATTR;
-        DTAR_user_opts.preserve = 1;
-        if (DTAR_global_rank == 0) {
+        archive_opts.flags |= ARCHIVE_EXTRACT_XATTR;
+        archive_opts.preserve = 1;
+        if (rank == 0) {
             MFU_LOG(MFU_LOG_INFO, "Creating archive with extended attributes");
         }
     }
 
-    DTAR_user_opts.chunk_size = opts_chunksize;
+    archive_opts.chunk_size = opts_chunksize;
 
-    /* init statistics */
-    DTAR_statistics.total_dirs  = 0;
-    DTAR_statistics.total_files = 0;
-    DTAR_statistics.total_links = 0;
-    DTAR_statistics.total_size  = 0;
-    DTAR_statistics.total_bytes_copied = 0;
-
-    if (DTAR_rank == 0) {
-        MFU_LOG(MFU_LOG_INFO, "Chunk size = %" PRIu64, DTAR_user_opts.chunk_size);
+    if (rank == 0) {
+        MFU_LOG(MFU_LOG_INFO, "Chunk size = %" PRIu64, archive_opts.chunk_size);
     }
 
     /* adjust pointers to start of paths */
     int numpaths = argc - optind;
-    const char** pathlist = &argv[optind];
+    const char** pathlist = (const char**) &argv[optind];
 
-    time(&(DTAR_statistics.time_started));
-    DTAR_statistics.wtime_started = MPI_Wtime();
     if (opts_create) {
-        /* check that input paths are valid
-         * (also set src_params and dest_param) */
-        DTAR_parse_path_args(numpaths, pathlist, opts_tarfile);
+        /* allocate space to record info for each source */
+        int num_src_params = numpaths;
+        size_t src_params_bytes = ((size_t) num_src_params) * sizeof(mfu_param_path);
+        mfu_param_path* src_params = (mfu_param_path*) MFU_MALLOC(src_params_bytes);
+    
+        /* process each source path */
+        mfu_param_path_set_all(numpaths, pathlist, src_params);
+    
+        /* standardize destination path */
+        mfu_param_path dest_param;
+        mfu_param_path_set(opts_tarfile, &dest_param);
+    
+        /* check that source and destination are okay */
+        int valid;
+        mfu_param_path_check_archive(num_src_params, src_params, dest_param, &valid);
 
         /* walk path to get stats info on all files */
         mfu_flist flist = mfu_flist_new();
         mfu_flist_walk_param_paths(num_src_params, src_params, 1, 0, flist);
 
         /* create the archive file */
-        mfu_flist_archive_create_libcircle(flist, opts_tarfile);
+        mfu_flist_archive_create(flist, opts_tarfile, &archive_opts);
 
         /* compress archive file */
         if (opts_compress) {
@@ -859,7 +282,7 @@ int main(int argc, const char** argv)
             strncpy(fname1, opts_tarfile, 50);
             strncpy(fname, opts_tarfile, 50);
             if ((stat(strcat(fname, ".bz2"), &st) == 0)) {
-                if (DTAR_rank == 0) {
+                if (rank == 0) {
                     printf("Output file already exists\n");
                 }
                 exit(0);
@@ -870,6 +293,10 @@ int main(int argc, const char** argv)
 
         /* free the file list */
         mfu_flist_free(&flist);
+
+        /* free paths */
+        mfu_param_path_free_all(num_src_params, src_params);
+        mfu_param_path_free(&dest_param);
     } else if (opts_extract) {
         char* tarfile = opts_tarfile;
         if (opts_compress) {
@@ -882,7 +309,7 @@ int main(int argc, const char** argv)
             printf("The file name is:%s %s %d", fname, fname_out, (int)len);
             struct stat st;
             if ((stat(fname_out, &st) == 0)) {
-                if (DTAR_rank == 0) {
+                if (rank == 0) {
                     printf("Output file already exists\n");
                 }
                 exit(0);
@@ -891,21 +318,16 @@ int main(int argc, const char** argv)
             remove(fname);
             tarfile = fname_out;
         }
-        extract_archive(tarfile, opts_verbose, DTAR_user_opts.flags);
+        mfu_flist_archive_extract(tarfile, opts_verbose, archive_opts.flags);
     } else {
-        if (DTAR_rank == 0) {
+        if (rank == 0) {
             MFU_LOG(MFU_LOG_ERR, "Neither creation or extraction is specified");
             DTAR_exit(EXIT_FAILURE);
         }
     }
 
-    DTAR_statistics.wtime_ended = MPI_Wtime();
-    time(&(DTAR_statistics.time_ended));
-
     /* free context */
     mfu_free(&opts_tarfile);
-    MFU_LOG(MFU_LOG_ERR, "Rank %d before epilogue\n", DTAR_rank);
-    DTAR_epilogue();
 
     DTAR_exit(EXIT_SUCCESS);
 }
